Updated the optimization pass manager to use LLVM 18.
Almost every member of `Pass` and `PassPipeliner` is deprecated, but the messages should point users in the right direction.
The expectation is that most, but not all, of a user’s code should still work without changes.

Added a `PassPipeline` type that corresponds to a sequence of `Passes`.
Renamed `PassPipeliner` to `PipelineManager`.

The functionality of `Pass.internalize` and `Pass.internalizeAll` may be impossible to implement with LLVM 18. These are the only tests that optimization pass related tests that fail and they have been removed.  
Anything that was deprecated before has now been removed.